Remove unnecessary conversions

+++ src/main/scala/eu/sim642/adventofcode2019/Day22.scala
@@ -36,7 +36,7 @@ object Day22 {
       case DealWithIncrement(n) =>
         val size = deck.size
         val nInv = NumberTheory.modInv(n, size)
-        Vector.tabulate(size)(i => deck(((nInv * i) % size).toInt))
+        Vector.tabulate(size)(i => deck((nInv * i) % size))
     }
 
     def shuffleFactoryOrder(techniques: Techniques, size: Long): Deck = {
